+++ src/detail/type_set.cpp
@@ -1,1 +1,1 @@
-#include "corofx/detail/type_set.hpp"
+#include "corofx/detail/type_set.hpp" // IWYU pragma: keep
