+++ src/trace.cpp
@@ -1,1 +1,1 @@
-#include "corofx/trace.hpp"
+#include "corofx/trace.hpp" // IWYU pragma: keep
